feat: refactor UI component organization and relocate setting widgets
Relocated numerous UI widgets from general utility packages to their respective feature-specific directories to improve project structure and modularity.

- **Refactoring**:
    - Moved `SettingsItemWidget.kt` and its associated components (e.g., `SettingsNavigationItemWidget`, `AutoLockInstaller`, `UninstallSystemAppWidget`) to `ui.page.main.settings.preferred`.
    - Moved `EditItemWidget.kt` and specific data widgets (e.g., `DataNameWidget`, `DataInstallModeWidget`) to `ui.page.main.settings.config.edit`.
    - Moved `ApplyItemWidget.kt` to `ui.page.main.settings.config.apply`.
    - Moved Miuix-specific widgets like `MiuixSettingsItemWidget.kt` and `MiuixEditItemWidget.kt` to their corresponding `preferred` and `edit` subdirectories.
- **Imports**: Updated extensive import references across `InstallChoiceContent`, `MiuixEditPage`, `PreferredPage`, `AboutPage`, and several other pages to reflect the new package locations.
- **Cleanup**: Removed redundant or unused imports in multiple UI files.
