+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/config/apply/NewApplyPage.kt
@@ -97,7 +97,6 @@ import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.page.main.widget.chip.Chip
 import com.rosan.installer.ui.page.main.widget.setting.AppBackButton
-import com.rosan.installer.ui.page.main.widget.setting.ApplyItemWidget
 import com.rosan.installer.ui.page.main.widget.setting.LabelWidget
 import com.rosan.installer.ui.theme.bottomShape
 import com.rosan.installer.ui.theme.getM3TopBarColor
