+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/config/edit/EditItemWidget.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 // Copyright (C) 2025-2026 InstallerX Revived contributors
-package com.rosan.installer.ui.page.main.widget.setting
+package com.rosan.installer.ui.page.main.settings.config.edit
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
@@ -46,8 +46,8 @@ import com.rosan.installer.domain.settings.model.InstallReason
 import com.rosan.installer.domain.settings.model.InstallerMode
 import com.rosan.installer.domain.settings.model.PackageSource
 import com.rosan.installer.ui.icons.AppIcons
-import com.rosan.installer.ui.page.main.settings.config.edit.EditViewAction
-import com.rosan.installer.ui.page.main.settings.config.edit.EditViewState
+import com.rosan.installer.ui.page.main.widget.setting.DropDownMenuWidget
+import com.rosan.installer.ui.page.main.widget.setting.SwitchWidget
 import com.rosan.installer.ui.util.isDhizukuActive
 import org.koin.compose.koinInject
