+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/config/edit/EditPage.kt
@@ -45,29 +45,6 @@ import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.page.main.widget.card.InfoTipCard
 import com.rosan.installer.ui.page.main.widget.dialog.UnsavedChangesDialog
 import com.rosan.installer.ui.page.main.widget.setting.AppBackButton
-import com.rosan.installer.ui.page.main.widget.setting.DataAllowAllRequestedPermissionsWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataAllowDowngradeWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataAllowTestOnlyWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataApkChooseAllWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataAuthorizerWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataAutoDeleteWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataBypassLowTargetSdkWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataCustomizeAuthorizerWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataDeclareInstallerWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataDescriptionWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataForAllUserWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataInstallModeWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataInstallReasonWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataInstallRequesterWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataManualDexoptWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataNameWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataPackageSourceWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataRequestUpdateOwnershipWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataShowToastWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataSplitChooseAllWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataUserWidget
-import com.rosan.installer.ui.page.main.widget.setting.DisplaySdkWidget
-import com.rosan.installer.ui.page.main.widget.setting.DisplaySizeWidget
 import com.rosan.installer.ui.page.main.widget.setting.LabelWidget
 import com.rosan.installer.ui.page.main.widget.util.EditEventCollector
 import com.rosan.installer.ui.theme.none
