+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/preferred/PreferredPage.kt
@@ -51,13 +51,7 @@ import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.navigation.Route
 import com.rosan.installer.ui.page.main.widget.card.InfoTipCard
 import com.rosan.installer.ui.page.main.widget.dialog.ErrorDisplayDialog
-import com.rosan.installer.ui.page.main.widget.setting.AutoLockInstaller
-import com.rosan.installer.ui.page.main.widget.setting.DefaultInstaller
-import com.rosan.installer.ui.page.main.widget.setting.DisableAdbVerify
-import com.rosan.installer.ui.page.main.widget.setting.IgnoreBatteryOptimizationSetting
 import com.rosan.installer.ui.page.main.widget.setting.LabelWidget
-import com.rosan.installer.ui.page.main.widget.setting.SettingsAboutItemWidget
-import com.rosan.installer.ui.page.main.widget.setting.SettingsNavigationItemWidget
 import com.rosan.installer.ui.page.main.widget.util.OnLifecycleEvent
 import com.rosan.installer.ui.theme.none
 import org.koin.androidx.compose.koinViewModel
