+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/preferred/subpage/installer/LegacyInstallerGlobalSettingsPage.kt
@@ -48,14 +48,14 @@ import com.rosan.installer.domain.settings.model.Authorizer
 import com.rosan.installer.domain.settings.model.InstallMode
 import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.navigation.LocalNavigator
+import com.rosan.installer.ui.page.main.settings.preferred.AutoClearNotificationTimeWidget
+import com.rosan.installer.ui.page.main.settings.preferred.DataAuthorizerWidget
+import com.rosan.installer.ui.page.main.settings.preferred.DataInstallModeWidget
+import com.rosan.installer.ui.page.main.settings.preferred.ManagedPackagesWidget
+import com.rosan.installer.ui.page.main.settings.preferred.ManagedUidsWidget
 import com.rosan.installer.ui.page.main.widget.setting.AppBackButton
-import com.rosan.installer.ui.page.main.widget.setting.AutoClearNotificationTimeWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataAuthorizerWidget
-import com.rosan.installer.ui.page.main.widget.setting.DataInstallModeWidget
 import com.rosan.installer.ui.page.main.widget.setting.IntNumberPickerWidget
 import com.rosan.installer.ui.page.main.widget.setting.LabelWidget
-import com.rosan.installer.ui.page.main.widget.setting.ManagedPackagesWidget
-import com.rosan.installer.ui.page.main.widget.setting.ManagedUidsWidget
 import com.rosan.installer.ui.page.main.widget.setting.SwitchWidget
 import com.rosan.installer.ui.theme.none
 import org.koin.androidx.compose.koinViewModel
