+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/preferred/subpage/lab/NewLabPage.kt
@@ -47,12 +47,12 @@ import com.rosan.installer.core.env.AppConfig
 import com.rosan.installer.domain.device.provider.DeviceCapabilityProvider
 import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.navigation.LocalNavigator
+import com.rosan.installer.ui.page.main.settings.preferred.LabHttpProfileWidget
+import com.rosan.installer.ui.page.main.settings.preferred.LabRootImplementationWidget
 import com.rosan.installer.ui.page.main.widget.card.InfoTipCard
 import com.rosan.installer.ui.page.main.widget.dialog.RootImplementationSelectionDialog
 import com.rosan.installer.ui.page.main.widget.setting.AppBackButton
 import com.rosan.installer.ui.page.main.widget.setting.IntNumberPickerWidget
-import com.rosan.installer.ui.page.main.widget.setting.LabHttpProfileWidget
-import com.rosan.installer.ui.page.main.widget.setting.LabRootImplementationWidget
 import com.rosan.installer.ui.page.main.widget.setting.SplicedColumnGroup
 import com.rosan.installer.ui.page.main.widget.setting.SwitchWidget
 import com.rosan.installer.ui.theme.getM3TopBarColor
