+++ app/src/main/java/com/rosan/installer/ui/page/main/settings/preferred/subpage/uninstaller/NewUninstallerGlobalSettingsPage.kt
@@ -45,15 +45,15 @@ import com.rosan.installer.R
 import com.rosan.installer.ui.activity.UninstallerActivity
 import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.navigation.LocalNavigator
+import com.rosan.installer.ui.page.main.settings.preferred.SettingsNavigationItemWidget
+import com.rosan.installer.ui.page.main.settings.preferred.UninstallForAllUsersWidget
+import com.rosan.installer.ui.page.main.settings.preferred.UninstallKeepDataWidget
+import com.rosan.installer.ui.page.main.settings.preferred.UninstallRequireBiometricAuthWidget
+import com.rosan.installer.ui.page.main.settings.preferred.UninstallSystemAppWidget
 import com.rosan.installer.ui.page.main.widget.card.InfoTipCard
 import com.rosan.installer.ui.page.main.widget.dialog.UninstallPackageDialog
 import com.rosan.installer.ui.page.main.widget.setting.AppBackButton
-import com.rosan.installer.ui.page.main.widget.setting.SettingsNavigationItemWidget
 import com.rosan.installer.ui.page.main.widget.setting.SplicedColumnGroup
-import com.rosan.installer.ui.page.main.widget.setting.UninstallForAllUsersWidget
-import com.rosan.installer.ui.page.main.widget.setting.UninstallKeepDataWidget
-import com.rosan.installer.ui.page.main.widget.setting.UninstallRequireBiometricAuthWidget
-import com.rosan.installer.ui.page.main.widget.setting.UninstallSystemAppWidget
 import com.rosan.installer.ui.theme.getM3TopBarColor
 import com.rosan.installer.ui.theme.installerHazeEffect
 import com.rosan.installer.ui.theme.none
