+++ app/src/main/java/com/rosan/installer/ui/page/miuix/settings/config/edit/MiuixEditPage.kt
@@ -38,30 +38,7 @@ import com.rosan.installer.ui.page.main.settings.config.edit.EditViewAction
 import com.rosan.installer.ui.page.main.settings.config.edit.EditViewEvent
 import com.rosan.installer.ui.page.main.settings.config.edit.EditViewModel
 import com.rosan.installer.ui.page.miuix.widgets.MiuixBackButton
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAllowAllRequestedPermissionsWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAllowDowngradeWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAllowTestOnlyWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataApkChooseAllWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAuthorizerWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAutoDeleteWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataBypassLowTargetSdkWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataCustomizeAuthorizerWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataDeclareInstallerWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataDescriptionWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataForAllUserWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataInstallModeWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataInstallRequesterWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataManualDexoptWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataNameWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataPackageSourceWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataSplitChooseAllWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataUserWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDisplaySdkWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDisplaySizeWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixInstallReasonWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixRequestUpdateOwnershipWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixSettingsTipCard
-import com.rosan.installer.ui.page.miuix.widgets.MiuixShowToastWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixUnsavedChangesDialog
 import com.rosan.installer.ui.theme.getMiuixAppBarColor
 import com.rosan.installer.ui.theme.installerHazeEffect
