+++ app/src/main/java/com/rosan/installer/ui/page/miuix/settings/preferred/MiuixPreferredItemWidget.kt
@@ -1,4 +1,4 @@
-package com.rosan.installer.ui.page.miuix.widgets
+package com.rosan.installer.ui.page.miuix.settings.preferred
 
 import androidx.annotation.StringRes
 import androidx.biometric.BiometricManager
@@ -52,6 +52,7 @@ import com.rosan.installer.domain.settings.model.SharedUid
 import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.uninstaller.UninstallerSettingsAction
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.uninstaller.UninstallerSettingsViewModel
+import com.rosan.installer.ui.page.miuix.widgets.MiuixSwitchWidget
 import com.rosan.installer.ui.theme.material.PaletteStyle
 import com.rosan.installer.ui.theme.material.ThemeColorSpec
 import com.rosan.installer.ui.theme.material.ThemeMode
