+++ app/src/main/java/com/rosan/installer/ui/page/miuix/settings/preferred/subpage/installer/MiuixInstallerGlobalSettingsPage.kt
@@ -49,13 +49,13 @@ import com.rosan.installer.ui.icons.AppIcons
 import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.installer.InstallerSettingsAction
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.installer.InstallerSettingsViewModel
-import com.rosan.installer.ui.page.miuix.widgets.MiuixAutoClearNotificationTimeWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixAutoClearNotificationTimeWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixDataAuthorizerWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixDataInstallModeWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixManagedPackagesWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixManagedUidsWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixBackButton
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataAuthorizerWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixDataInstallModeWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixIntNumberPickerWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixManagedPackagesWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixManagedUidsWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixSwitchWidget
 import com.rosan.installer.ui.theme.getMiuixAppBarColor
 import com.rosan.installer.ui.theme.installerHazeEffect
