+++ app/src/main/java/com/rosan/installer/ui/page/miuix/settings/preferred/subpage/theme/MiuixThemeSettingsPage.kt
@@ -46,14 +46,14 @@ import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.theme.ThemeSettingsAction
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.theme.ThemeSettingsViewModel
 import com.rosan.installer.ui.page.main.widget.card.ColorSwatchPreview
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixColorSpecWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixPaletteStyleWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixThemeEngineWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixThemeModeWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixBackButton
 import com.rosan.installer.ui.page.miuix.widgets.MiuixBlurWarningDialog
-import com.rosan.installer.ui.page.miuix.widgets.MiuixColorSpecWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixHideLauncherIconWarningDialog
-import com.rosan.installer.ui.page.miuix.widgets.MiuixPaletteStyleWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixSwitchWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixThemeEngineWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixThemeModeWidget
 import com.rosan.installer.ui.theme.getMiuixAppBarColor
 import com.rosan.installer.ui.theme.installerHazeEffect
 import com.rosan.installer.ui.theme.rememberMiuixHazeStyle
