+++ app/src/main/java/com/rosan/installer/ui/page/miuix/settings/preferred/subpage/uninstaller/MiuixUninstallerGlobalSettingsPage.kt
@@ -34,14 +34,14 @@ import com.rosan.installer.ui.activity.UninstallerActivity
 import com.rosan.installer.ui.navigation.LocalNavigator
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.uninstaller.UninstallerSettingsEvent
 import com.rosan.installer.ui.page.main.settings.preferred.subpage.uninstaller.UninstallerSettingsViewModel
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixNavigationItemWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixUninstallForAllUsersWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixUninstallKeepDataWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixUninstallRequireBiometricAuthWidget
+import com.rosan.installer.ui.page.miuix.settings.preferred.MiuixUninstallSystemAppWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixBackButton
-import com.rosan.installer.ui.page.miuix.widgets.MiuixNavigationItemWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixSettingsTipCard
-import com.rosan.installer.ui.page.miuix.widgets.MiuixUninstallForAllUsersWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixUninstallKeepDataWidget
 import com.rosan.installer.ui.page.miuix.widgets.MiuixUninstallPackageDialog
-import com.rosan.installer.ui.page.miuix.widgets.MiuixUninstallRequireBiometricAuthWidget
-import com.rosan.installer.ui.page.miuix.widgets.MiuixUninstallSystemAppWidget
 import com.rosan.installer.ui.theme.getMiuixAppBarColor
 import com.rosan.installer.ui.theme.installerHazeEffect
 import com.rosan.installer.ui.theme.rememberMiuixHazeStyle
